Add calculateDefaultPosition() method for DummyMotor
If a motor has a min limit of 2.1 for example and the currentPosition is
initially zero when starting new gda server, the motor cannot move
because it will forever be outside the limits. This fixes it by
adjusting the default position when creating a motor position file to be
within the limits.

Change-Id: Ib1e88c78fba135fdd2ad8d6b2960ad6b8a71a92b