DAQ-5237 - Update INexusDeviceService factory methods to allow detectors to work per scan monitor if supported
If a detector is added as monitor per scan, it crashes with
NexusScanDataWriter because the nexus device wrapper are not designed to
work per scan. However, with NexusDataWriter it works fine because it
treats it as a scannable and does getPosition() on it.

This fix mirrors NexusDataWriter as it treats all per scan monitor as a
scannable and therefore wraps it with the DefaultScannableNexusDevice
which uses getPosition() to save data, the same as NexusDataWriter.

Change-Id: I9557de7d804064fe2d222f8d9a2bc28bdb17aafa