I08-1 Add documentation to the PolarisationSubmitScanSection
In preparation for a change that is only deployed at the beamline, I had to read the classes related to the Polarisation scans. I realised, they should have more documentation to understand what this class does.

The change includes new Javadocs and a small refactor of a few methods that should not affect functionality.

Change-Id: I3d9dc67dca77401307c727be061a346909bda72d