I08-1 Add new Polarisation setting
This is an old change that was on I08-1 deployment. They asked me to add a new "Polarisation setting" for a linear arbitrary angle (Mn), where depending on the degree selected, there is a different energy target value (this is in the polarisation script). This change has been there for a long time but now I want to push it to master as they want this to be a permanent change.

Change-Id: I49418b5771e3b98e02e1d564258969544d171d9f