P60-6 - Adopt suitable class for scannable that holds a constant value
P60 will be using radioactive sources, Mg K-Alpha and Al K-Alpha, for
their excitation energy and values won't change. The existing framework
that I09 uses needs these excitation energy to be scannables. Therefore,
a class needs to be used where a scannable can take on a constant value
and cannot be moved.

These changes allow for additional configuration to be applied to allow
for this requirement to be met.

Change-Id: I133ddbb8e7f6a51d79d868e38d8f36e4ed0fa7a7