P60-8 - Check to see if sequence file loaded is compatible with defined configuration
When loading a sequence file, it may have different configuration as it
may have been created from a different beamline. This focuses on being
able to still load the file by converting it to to the defined
configuration so there are no crashes.

Other changes include
- Better handling of region validation by catching
IllegalArgumentExceptions if the region had an invalid lens or pass
energy.
- Updated lens mode and pass energy to use CCombo so that invalid values
can be displayed why displaying the error message to user to change
them.
- Removed hard coded lens modes to be removed and instead comes from
configuration.
- Update AnalyserEnergyRangeConfiguration to read from file that is
comma separated rather than tab separated.

Change-Id: Ic28719ae1d292929ed6e226d5f38e58205f2c331