I14 Modify "Clear submitted jobs queue" action
I added a feature to clear only the submission queue. However, having both clearing queue functionalities in the same method was a bit confusing when to determine the purpose of each one.

There is now some code duplication from separating the logic into two methods, but I believe keeping them separate makes the implementation more readable and easier to understand.

Additionally, if the goal is to only clear the submitted jobs, I don’t think it's necessary to show a popup asking whether we want to terminate active scans. So, I’ve removed the terminateActiveScans method. This method was originally intended to address "ghost" scans (those that had a "RUNNING" status but weren’t actually running), but it wasn’t effective in handling those scans. Furthermore, terminating active scans should not be part of the action that clears the submitted jobs.

UI changes:

- Previously, I was unable to set an image for the "Clear Queue..." MenuAction (that contains both clearing queue actions). So, the queue action displayed as "Clear queue..." instead of showing the icon. I’ve now added an icon for this action. However, we can see the options by both selecting the icon and the dropdown arrow next to it.

Change-Id: I0e9a83d63bf3871ad870773f10a1a4c60eaaf0dc