I14-1097 Add "Line to track" parameter to Standards Scan script
For post processing, I14 needs to know the element/edge.

As there is already a dropdown box for this, that is used for generating
the energies, and the element/edge for the postprocessing is going to be
the same. I've just created a new LineToTrack object based on the combo
text. There is an extra "a" added at the end, as all post processing
scripts look for "Ka" instead of "K" in the string.

Change-Id: I2dbbee72654450e3992fc4b5f0484923d0a95131