I14-1098 Add feature to calculate and record centre of rotation
Add new section in the Mapping GUI, only visible for the Tomography
scans GUI.

It creates two columns to record x,y,z and rot positions. Then, these
values are used to calculate the centre of rotation.

Process description: user finds the ideal x, y and z positions
(crosshairs) for a determined rotation. Then, moves the rotation stage
180 degrees, finds the ideal positions for that rotation. And, from the
values of the two columns, it calculates the centre of rotation, and
then they can click a button to move the stages to these positions.

Change-Id: I8930ba860c8c778820ee8608fc136ea1ed684dc3