I09-596 - Further improve reconnection for element set
If the server is started before it is connected to the analyser, the
analyser would not work because it wouldn't be configured. The configure
method sets the valid pass energies, detector modes, angular values, and
element set modes. If an analyserscan is attempted, it would fail
stating invalid detectorMode because the valid list is empty. If the
client was started, it would fail to recover from disconnect because the
valid values for elementSet would be empty.

This fixes it by checking that the VGScientaAnalyser and
VGScientaController are configured when setting a region or when setting
one of the listed parameters that it is configured, if not configure it.

This also improves the client to listen to the analyser_element_set
scannable which is injected to the client.

EpicsSimplePositioner is further improved as it can now recover from
disconnections and also has a setting to check if it can be moved from
within GDA.

Change-Id: I67476436f8583db8272b375346b8535686cc7e5a