I14-1088 Add sample transfer control panel
Composite that allows users to set whether a hotel holder contains a
sample, and which is the name of the sample. When they are happy with
the changes, they can press the "Save settings" button and start moving
the samples. (They can also press the "Enable changes" button if they
made a mistake)

The interface allows users to move the samples between the holder and
the spigot and updates the state and name fields. In the case of the
spigot, there is a combo viewer to select which sample holder the sample
is going to be moved to.

The activate/deactivate buttons functionality needs to be implemented,
as well as the actual process if moving the samples.

Change-Id: I1608a3559df989736b2a88c70fedead402dd139d