Add runAsJython method to run java in JythonServerThread
In order to support per-user authentication for blueAPI, access to the
client aware components of the GDA server (the JythonServerThread
sub-classes) need to be restricted (to prevent inappropriate access).

The existing uses of these threads by external (to the JythonServer)
packages for running non-jython code still need to be supported so the
new runAsJython method allows Java code to be run in a Jython thread
without client code needing to create the Jython threads themselves.

Change-Id: Idd5f8ab252b276460f3c91324d1ce7a89b313276