I07-543:  Disable processing for multi dimensional scans
SwmrHdfDatasetProviderProcessor has not been designed with multi
dimensional (i.e. multiple motor grid scans) in mind, it cannot read the
dataset.  See change 43588 and the ticket I07-543 for my attempt to add
this and the reasons it can't currently be done.  This change will
disable all the processors so the scan will still run.

Change-Id: I9363d68d8914e0327a122417185e272d0089a682