I20-655 : Fix for XesSpectrometer scannable offsets
GDA offsets are now applied correctly to analyser motor moves when using
the 'deferred move' option. Normally, the addition of the offset to the
demand value is handled internally by each ScannableMotor, but
PVScannables are used for deferred moves and these do not include an
offset. Therefore, the demand values need to be adjusted to include the
offset before they are passed to the PVScannables.

Change-Id: I1c4c3db442681ba676b608349ecb19f74731ec1e