I18-660 : Add links in Nexus file to point to Xspress3 Odin MCA data
The MCA data written by the Odin hdf writer is spread across several Hdf
files (4 channels per file). So instead of a single link to the MCA data
(as was done for areadetector based Xspress3), a link to data for each
channel of the detector is added to the Nexus file. The link naming
convention follows that used by mapping scans (i.e. raw_mca0, raw_mca1
etc)

Change-Id: I5b97c3ac5666ff4e5680a22f555301fddb3b63cc