Update 'live background' related classes to use OSGI services
The Spring services used for plotting live background are sometimes not
available to the client - depending which java packages are included in
class path scanning. This can can problems in the client - e.g. live
plotting not working, or the client failing to start up.
The changes here make live background related classes use OSGI services
so that these features work correctly without needing the client to
include the correct packages in the class path scanning setup.
BackgroundStateHelper will now attempt to connect to the live stream
several times if exceptions are thrown - this helps avoid problems with
the camera CalibratedAxisProvider dataset being unavailable due to race
conditions.

Change-Id: I1bc634b0c4d92e5e52b4eee01803a8ea7998a110