I14-1089 Refactor EXAFS GUI
This refactor removes the XANES parameters from the EXAFS gui. As the element/edge combo generates a list of XANES energies that are not true to EXAFS.

The energies are generated in a Jython script after submitting the scan based on the element/edge selected.

The submit section now doesn't extend the XANES submit section and doesn't display the XANES parameters section in the Mapping GUI.

The dedicated EXAFS mapping section is now two sections, one with parameters related to the energy and K, and the other one related to the exposure time for K values.

Change-Id: Ie9ed9b49babeea58765c62e4b6ac209ef33797f9