I14-1102 Update Element/edge combo selection handler
In the StandardsView, the Element/edge combo viewer was updating the scan path with a list of energies. These energies do not represent an EXAFS scan, but the user should still select an element/edge to select the edge energy to generate the list of energies in the script.

In the UI, the element/edge combo handler should display the associated energy.

This change refactors the class to override some of the methods to create the composite and set the scan parameters.

Change-Id: I9d000e791e529acd58039b907fb8385e24546cf1