I14-1114 Add data driven scan type
This change adds a new scan type within the Mapping GUI to run 'data
driven scans'. This type of scan consists of two stages. In the first
stage, there is an initial full XANES scan with 5 energies, with all y
positions. The results of this scan are sent to be post processed.

The post processing job will generate a JSON file stored in gda_var,
with a dictionary of energies as keys and y positions as values. The
second stage of the scan consists of submitting a scan with these
defined energies and y positions.

This change adds two submit buttons to trigger both processes. The user
must check that the JSON file exists before submitting the second scan.

The logic for these scans are in Jython scripts that are located in
dls_sw/i14/scripts/...

In this change, I have refactored the XANES and Sparse XANES section to
split the creation of controls methods so that subclasses can override
their functionality.

Change-Id: If0f0b56eb5624e1c1574b1c084f6a58cbfd6e9dc