I14-1088 Refactor Sample Transfer UI
The previous GUI was a bit complex, as the user had to manage the state of the holder positions while managing the transfer of samples.
With this refactor, the dialogue gets simplified, and now the different stages in the process are in different tabs. This change only adds functionality to the 'sample selection' tab, it also creates a second tab but it is empty at the moment.

In the sample selection composite, there are three columns, one with the title of the holder position, the second column contains a check box to indicate whether the holder is being occupied or is empty (an red/green icon indicates this state) and a third column for the user to write the sample associated to the holder position.

Every time a check box is being selected or a name is being writter, the system writes this information in a JSON file. When the dialog is opened, there is a user confirmation message asking the user whether they want to load the state from file or not.


Change-Id: Ibd2f640f1ecdadcfab160c40be62b2fca18eec0f