I14-1088 Add composite for transfering samples
This is a change for the second tab for managing the transfer of
samples. It also refactors the first tab slightly so both composites
maintain the same map of HolderState objects.

Any changes to a holder state, either in the selection composite and the
transfer composite are saved immediately to a file to ensure
persistence.

HolderTransferComposite provides functionality for transferring samples
between holders and the spigot. It uses the same shared state map and
updates the UI accordingly to reflect sample transfers.

Transfer buttons are enabled/disabled on the current state, and after
every transfer, the update state is saved to file.

Change-Id: I36b5baad47bd928c139e0576c2cfed7b9002fb36